refactor: move CVLData struct to balance.rs
