+++ core/credit/src/ledger/balance.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 use core_money::{Satoshis, UsdCents};
 
-use crate::{primitives::CVLData, CVLPct};
+use crate::CVLPct;
 
 #[cfg(not(test))]
 #[derive(Debug, Default, Copy, Clone, Serialize, Deserialize)]
@@ -133,6 +133,27 @@ impl CreditFacilityBalanceSummary {
         Self {
             disbursed: self.disbursed + disbursal,
             ..self
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+struct CVLData {
+    amount: UsdCents,
+    collateral: Satoshis,
+}
+
+impl CVLData {
+    fn new(collateral: Satoshis, amount: UsdCents) -> Self {
+        Self { collateral, amount }
+    }
+
+    fn cvl(&self, price: PriceOfOneBTC) -> CVLPct {
+        let collateral_value = price.sats_to_cents_round_down(self.collateral);
+        if collateral_value == UsdCents::ZERO {
+            CVLPct::ZERO
+        } else {
+            CVLPct::from_loan_amounts(collateral_value, self.amount)
         }
     }
 }
